+++ src/core/constants.ts
@@ -1,0 +1,5 @@
+export const ERROR_MESSAGES = {
+  tusClientHasNotFounded: 'No TusClient set, use TusClientProvider to set one',
+  tusIsNotSupported:
+    'This browser does not support uploads. Please use a modern browser instead.',
+};
