+++ src/tests/useTus.test.tsx
@@ -1,13 +1,11 @@
 import { FC } from 'react';
 import { Upload } from 'tus-js-client';
 import { renderHook, act } from '@testing-library/react-hooks';
-import {
-  ERROR_MESSAGES,
-  TusClientProvider,
-  useTusClientState,
-} from '../TusClientProvider';
+import { TusClientProvider } from '../TusClientProvider';
 import { useTus } from '../useTus';
 import { getBlob } from './utils/getBlob';
+import { ERROR_MESSAGES } from '../core/constants';
+import { useTusClientState } from '../core/contexts';
 
 const getDefaultOptions: () => Upload['options'] = () => ({
   endpoint: 'http://tus.io/uploads',
